Added new main task, "4 bit parity"
There are 16 possible 4-bit inputs (from 0000 to 1111 in binary). The parity is 1 if the total number of 1-bits is odd, and 0 if even. That’s effectively a 4-bit generalization of XOR (the “sum” mod 2). We store each 4-bit pattern in a Matrix(1, 4) and compute the number of ones to determine the target label (0 or 1).